feat(share article): Users should be able to share articles on different platforms (#19)
This commit delivers the 'share an article' feature
The platforms we can share to are; facebook, twitter, and email

[Delivers #162163237]

feat(like and dislike): A user should be able to like and dislike articles. (#20)

This commit ensures that a user can like or dislike articles.

[Delivers #162163235]

ch(Slack/Heroku Intergrations): Setup slack intergration with Travis and Heroku (#9)

This commit delivers a chore for slack intergration with Travis CI
and Heroku
Slack receives notifications directly from Travis CI after every  build
The same for Heroku

[Delivers #163015208]

mdm